[exporterhelper] Enable Persistent Queue feature by default
