+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 
 - Move `service.mapResolver` to `confmap.Resolver` (#5444)
 - Add `linux-arm` architecture to cross build tests in CI (#5472)
+- Enable Persistent Queue by default. (#5457)
 
 ### 🧰 Bug fixes 🧰
